pyproject.toml
[project]
name = "herculesai-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.2",
    "openai>=1.42.0",
    "pydantic>=2.8.2",
    "python-fasthtml>=0.4.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
ignore = ["F403", "F405"]
